requirements.txt
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
Pillow==9.5.0
requests==2.31.0 